finish basic functionalities and refactoring
