+++ src/App.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
 
-import Layout from './components/Layout/Layout'
+import Layout from './hoc/Layout/Layout'
 import BurgerBuilder from './containers/BurgerBuilder/BurgerBuilder'
 
 class App extends Component {
