+++ src/components/UI/Modal/Modal.js
@@ -1,18 +1,31 @@
-import React from 'react';
+import React, { Component } from 'react';
 import classes from './Modal.css';
-import Aux from '../../../hoc/Aux';
+import Aux from '../../../hoc/Aux/Aux';
 import Backdorp from '../Backdrop/Backdrop'
 
-const modal= (props) => (
-    <Aux>
-        <Backdorp show={props.show} clicked={props.modalClosed}/>
-        <div
-            className={classes.Modal}
-            style={{transform: props.show ? 'translateY(0)' : 'translate(-100vh)',
-            opacity: props.show ? '1' : '0'}}>
-            {props.children}
-        </div>
-    </Aux>
-);
+class  Modal extends Component {
 
-export default modal;
+    shouldComponentUpdate(nextProps, nextState) {
+        return nextProps.show !== this.props.show;
+    }
+
+    componentDidUpdate() {
+        console.log('[Modal didUpdate]');
+    }
+
+    render(){
+        return(
+            <Aux>
+                <Backdorp show={this.props.show} clicked={this.props.modalClosed}/>
+                <div
+                    className={classes.Modal}
+                    style={{transform: this.props.show ? 'translateY(0)' : 'translate(-100vh)',
+                        opacity: this.props.show ? '1' : '0'}}>
+                    {this.props.children}
+                </div>
+            </Aux>
+        );
+    };
+};
+
+export default Modal;
